{{ cookiecutter.project_slug }}/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "app"
version = "0.1.0"
description = "cookiecutter template for fastapi"
license = "MIT"
authors = ["caerulescens <caerulescens.github@proton.me>"]
maintainers = ["caerulescens <caerulescens.github@proton.me>"]
readme = "README.md"
packages = [{include = "{{ cookiecutter.package_slug }}", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
uvloop = "^0.19.0"
orjson = "^3.9.10"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
fastapi = "^0.105.0"
fastapi-users = {extras = ["oauth", "sqlalchemy"], version = "^12.1.2"}
pydantic = "^2.5.2"
pydantic-settings = "^2.1.0"
sqlmodel = "^0.0.14"
sqlalchemy = "^2.0.23"
alembic = "^1.13.0"
asyncpg = "^0.29.0"
loguru = "^0.7.2"
prometheus-fastapi-instrumentator = "^6.1.0"
opentelemetry-instrumentation-fastapi = "^0.42b0"
opentelemetry-exporter-otlp = "^1.21.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
pytest-sugar = "^0.9.7"
pytest-async-sqlalchemy = "^0.2.0"
pytest-postgresql = "^5.0.0"
psycopg = {extras = ["binary"], version = "^3.1.11"}
aioresponses = "^0.7.4"
coverage = {extras = ["toml"], version = "^7.3.1"}
tox = "^4.11.4"
black = "^23.9.1"
ruff = "^0.1.7"
mypy = "^1.7.1"
locust = "^2.19.1"
sphinx = "^7.2.6"
sphinx_rtd_theme = "2.0.0"
myst-parser = "^2.0.0"
pydocstyle = "^6.3.0"
cython = "^3.0.2"
pre-commit = "^3.4.0"

[tool.pytest.ini_options]
minversion = "7.1.1"
addopts = "-ra -q"
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
concurrency = ["gevent"]
dynamic_context = "test_function"

[tool.coverage.report]
fail_under = 50

[tool.coverage.html]
directory = "coverage_html_report"

[tool.black]
line-length = 88
target-version = ["py310", "py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py310"
src = ["src", "tests", "docs"]
select = ["F", "E", "W", "C90", "S", "I", "N", "D", "UP"]
ignore = ["D100", "D104", "D106"]

[tool.ruff.pydocstyle]
convention = "pep257"

[tool.mypy]
python_version = "3.11"
files = ["src", "tests", "docs"]
plugins = ["pydantic.mypy"]
follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true
